backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML & Image Processing
tensorflow==2.15.0
Pillow==10.1.0
numpy==1.26.2
scikit-learn==1.3.2
opencv-python-headless==4.8.1.78

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.23.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# PDF Generation
fpdf2==2.7.6

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0
pydantic[email]==2.5.2

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2

# Data Augmentation 
albumentations==1.3.1

# CORS
starlette==0.27.0

# Testing
pytest==7.4.3
httpx==0.25.2
